webservice/requirements.txt
tensorflow==2.12.0
pandas==2.0.2
numpy==1.23.5
Pillow==9.5.0
Flask==2.3.2
requests==2.31.0
waitress==2.1.2
mlflow==2.4.1
prefect   #==2.10.0
prefect-email==0.2.2